omics_oracle/query_manager.py
# omics_oracle/query_manager.py

from .spoke_wrapper import SpokeWrapper
from typing import Dict, Any
from .logger import setup_logger

class QueryManager:
    def __init__(self, spoke_wrapper: SpokeWrapper):
        self.spoke = spoke_wrapper
        self.logger = setup_logger(__name__)

    def process_query(self, user_query: str) -> Dict[str, Any]:
        """
        Process a user query through the Spoke API.

        Args:
            user_query (str): The natural language query from the user.

        Returns:
            Dict[str, Any]: A dictionary containing the processed results.
        """
        self.logger.info(f"Processing user query: {user_query}")
        try:
            # Step 1: Convert user query to AQL query (placeholder)
            aql_query = self.convert_to_aql(user_query)
            self.logger.debug(f"AQL query generated: {aql_query}")

            # Step 2: Execute the AQL query on the Spoke knowledge graph
            if aql_query:
                self.logger.debug(f"Executing AQL query on SPOKE: {aql_query}")
                spoke_results = self.spoke.execute_aql(aql_query)
                self.logger.debug(f"Results from SPOKE: {spoke_results}")
            else:
                spoke_results = []
                self.logger.warning("No AQL query generated. Skipping Spoke query execution.")

            # Step 3: Interpret the results (placeholder)
            interpretation = self.interpret_results(spoke_results, user_query)
            self.logger.debug(f"Interpretation: {interpretation}")

            return {
                "original_query": user_query,
                "aql_query": aql_query,
                "spoke_results": spoke_results,
                "interpretation": interpretation
            }
        except Exception as e:
            self.logger.error(f"Error processing query: {e}")
            raise

    def convert_to_aql(self, user_query: str) -> str:
        """
        Convert the user query to an AQL query.
        This is a placeholder and should be implemented with appropriate logic.
        """
        # Placeholder implementation
        self.logger.debug(f"Converting user query to AQL: {user_query}")
        # For now, we'll just return a simple AQL query
        return "FOR doc IN collection FILTER doc.type == 'biomedical' RETURN doc"

    def interpret_results(self, spoke_results: list, original_query: str) -> str:
        """
        Interpret the SPOKE results in the context of the original query.
        This is a placeholder and should be implemented with appropriate logic.
        """
        # Placeholder implementation
        self.logger.debug(f"Interpreting results for query: {original_query}")
        return f"Found {len(spoke_results)} results related to the query."